+++ src/main/java/com/tcode/moviebase/Services/ReportMessageService.java
@@ -1,10 +1,14 @@
 package com.tcode.moviebase.Services;
 
 
+import com.tcode.moviebase.Dtos.ReportMessageDto;
 import com.tcode.moviebase.Entities.ReportMessage;
+import com.tcode.moviebase.Exceptions.MessageDoesntExistsException;
 import com.tcode.moviebase.Repositories.MessageRepository;
 import com.tcode.moviebase.Repositories.ReportMessageRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +22,7 @@ public class ReportMessageService {
 
 
     public ReportMessage report(Long messageId, String reason) {
-        var message = messageRepository.findById(messageId).orElse(null);
+        var message = messageRepository.findById(messageId).orElseThrow( () -> new MessageDoesntExistsException("Message doesn't exist"));
         var reportMessage = new ReportMessage();
         reportMessage.setMessage(message);
         reportMessage.setReason(reason);
@@ -26,25 +30,37 @@ public class ReportMessageService {
         return reportMessageRepository.save(reportMessage);
     }
 
-    public List<ReportMessage> getAllReports() {
-        return reportMessageRepository.findAll();
+    public Page<ReportMessageDto> getAllReports(Pageable pageable) {
+        return reportMessageRepository.findAll(pageable).map(this::reportMessageToReportMessageDto);
     }
 
     @Transactional
     public void approveReport(Long reportId) {
-        var report = reportMessageRepository.findById(reportId).orElse(null);
-        if (report != null && report.getMessage() != null) {
-            reportMessageRepository.delete(report);
-            messageRepository.delete(report.getMessage());
-        }
+        var report = reportMessageRepository.findById(reportId).orElseThrow(() -> new MessageDoesntExistsException("Report doesn't exist"));
+
+        reportMessageRepository.delete(report);
+        messageRepository.delete(report.getMessage());
+
     }
 
     public void rejectReport(Long reportId) {
-        reportMessageRepository.deleteById(reportId);
+        if (reportMessageRepository.findById(reportId).isPresent()) {
+            reportMessageRepository.deleteById(reportId);
+        } else {
+            throw new MessageDoesntExistsException("Report doesn't exist");
+        }
     }
 
 
     public boolean existsById(Long reportId) {
         return reportMessageRepository.existsById(reportId);
+    }
+
+    private ReportMessageDto reportMessageToReportMessageDto(ReportMessage reportMessage) {
+        var dto = new ReportMessageDto();
+        dto.setReportId(reportMessage.getId());
+        dto.setMessage(reportMessage.getMessage().getContent());
+        dto.setReason(reportMessage.getReason());
+        return dto;
     }
 }
